Folder structure refactoring, seperated into layered architecture
