+++ OatMilk.Backend.Api/Controllers/Security/JWTClaimTypes.cs
@@ -1,0 +1,7 @@
+﻿namespace OatMilk.Backend.Api.Controllers.Security
+{
+    public static class JWTClaimTypes
+    {
+        public static string UserId = "UserId";
+    }
+}
