+++ OatMilk.Backend.Api/Controllers/Security/Requirements/UserAuthorizationRequirement.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNetCore.Authorization;
+
+namespace OatMilk.Backend.Api.Controllers.Security.Requirements
+{
+    public class UserAuthorizationRequirement : IAuthorizationRequirement
+    {
+    }
+}
