--- OatMilk.Backend.Api/Data/Models/Enums/DurationType.cs
@@ -1,9 +1,0 @@
-﻿namespace OatMilk.Backend.Api.Data.Models.Enums
-{
-    public enum DurationType
-    {
-        Instant, // Effect will be applied periodically for the duration
-        Duration, // Effect will be applied instantly, once
-        Infinite // Effect will be applied periodically for infinite duration
-    }
-}
