--- OatMilk.Backend.Api/Data/Models/Enums/ErrorType.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace OatMilk.Backend.Api.Data.Models.Enums
-{
-    public enum ErrorType
-    {
-        Error,
-        ArgumentError
-    }
-}
