--- OatMilk.Backend.Api/Data/Models/Enums/ModifierOperation.cs
@@ -1,10 +1,0 @@
-﻿namespace OatMilk.Backend.Api.Data.Models.Enums
-{
-    public enum ModifierOperation
-    {
-        Add = 0,
-        Multiply = 1,
-        Divide = 2,
-        Override = 3
-    }
-}
