--- OatMilk.Backend.Api/Data/Models/Requests/AbilityRequest.cs
@@ -1,7 +1,0 @@
-﻿namespace OatMilk.Backend.Api.Data.Models.Requests
-{
-    public class AbilityRequest
-    {
-        
-    }
-}
