--- OatMilk.Backend.Api/Data/Models/Requests/UserLoginRequest.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace OatMilk.Backend.Api.Data.Models.Requests
-{
-    public class UserLoginRequest
-    {
-        [Required]
-        [EmailAddress(ErrorMessage = "Not a valid email address.")]
-        public string Email { get; set; }
-
-        [Required]
-        public string Password { get; set; }
-    }
-}
