--- OatMilk.Backend.Api/Data/Repositories/Abstraction/IRepository.cs
@@ -1,16 +1,0 @@
-﻿using System.Linq;
-using System.Threading.Tasks;
-
-namespace OatMilk.Backend.Api.Data.Repositories.Abstraction
-{
-    public interface IRepository<TEntity> where TEntity :class
-    {
-        Task<TEntity> FindAsync(params object[] keyValues);
-        
-        IQueryable<TEntity> Get();
-
-        void Add(TEntity entity);
-        
-        Task SaveAsync();
-    }
-}
