+++ OatMilk.Backend.Api/Repositories/Abstraction/IRepository.cs
@@ -1,0 +1,16 @@
+﻿using System.Linq;
+using System.Threading.Tasks;
+
+namespace OatMilk.Backend.Api.Repositories.Abstraction
+{
+    public interface IRepository<TEntity> where TEntity :class
+    {
+        Task<TEntity> FindAsync(params object[] keyValues);
+        
+        IQueryable<TEntity> Get();
+
+        void Add(TEntity entity);
+        
+        Task SaveAsync();
+    }
+}
