--- OatMilk.Backend.Api/Security/JWTClaimTypes.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace OatMilk.Backend.Api.Security
-{
-    public static class JWTClaimTypes
-    {
-        public static string UserId = "UserId";
-    }
-}
