--- OatMilk.Backend.Api/Security/Requirements/UserAuthorizationRequirement.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.AspNetCore.Authorization;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace OatMilk.Backend.Api.Security.Requirements
-{
-    public class UserAuthorizationRequirement : IAuthorizationRequirement
-    {
-    }
-}
