+++ OatMilk.Backend.Api/Services/Models/Enums/DurationType.cs
@@ -1,0 +1,9 @@
+﻿namespace OatMilk.Backend.Api.Services.Models.Enums
+{
+    public enum DurationType
+    {
+        Instant, // Effect will be applied periodically for the duration
+        Duration, // Effect will be applied instantly, once
+        Infinite // Effect will be applied periodically for infinite duration
+    }
+}
