+++ OatMilk.Backend.Api/Services/Models/Enums/ErrorType.cs
@@ -1,0 +1,8 @@
+﻿namespace OatMilk.Backend.Api.Services.Models.Enums
+{
+    public enum ErrorType
+    {
+        Error,
+        ArgumentError
+    }
+}
