+++ OatMilk.Backend.Api/Services/Models/Enums/ModifierOperation.cs
@@ -1,0 +1,10 @@
+﻿namespace OatMilk.Backend.Api.Services.Models.Enums
+{
+    public enum ModifierOperation
+    {
+        Add = 0,
+        Multiply = 1,
+        Divide = 2,
+        Override = 3
+    }
+}
