+++ OatMilk.Backend.Api/Services/Models/Requests/AbilityRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace OatMilk.Backend.Api.Services.Models.Requests
+{
+    public class AbilityRequest
+    {
+        
+    }
+}
