+++ OatMilk.Backend.Api/Services/UserService.cs
@@ -13,37 +13,6 @@ using OatMilk.Backend.Api.Data.Repositories.Abstraction;
 
 namespace OatMilk.Backend.Api.Services
 {
-    public interface IUserService
-    {
-        /// <summary>
-        /// Get user profile.
-        /// </summary>
-        /// <returns></returns>
-        UserResponse GetUser(Guid userId);
-
-        /// <summary>
-        /// Login to a user account.
-        /// </summary>
-        /// <param name="request"></param>
-        /// <returns>Auth JWT to use for other endpoints.</returns>
-        AuthTokenResponse Login(UserLoginRequest request);
-
-        /// <summary>
-        /// Register a new user account.
-        /// Will login the user as well.
-        /// </summary>
-        /// <param name="request"></param>
-        /// <returns>Auth JWT to use for other endpoints.</returns>
-        Task<AuthTokenResponse> Register(UserRegisterRequest request);
-
-        /// <summary>
-        /// Check if a user of id <paramref name="userId"/> exists.
-        /// </summary>
-        /// <param name="userId">Id of user</param>
-        /// <returns>True if user exists, false otherwise.</returns>
-        bool UserExistsById(Guid userId);
-    }
-
     public class UserService : IUserService
     {
         private readonly IConfiguration _configuration;
